Relatorio_GitLab/requirements.txt
streamlit
pandas
plotly_express
datetime
requests
python-dateutil